feat(functions/v2): use FF testing layer in audit log sample (#2647)
* updated audit log v2 sample to new FF testing layer

* fixed lint
