+++ functions/v2/helloAuditLog/.gcloudignore
@@ -1,0 +1,3 @@
+node_modules/
+test/
+package-lock.json
