+++ functions/v2/helloAuditLog/package.json
@@ -12,16 +12,14 @@
     "node": ">=12.0.0"
   },
   "scripts": {
-    "test": "mocha test/index.test.js"
+    "test": "mocha test/*.test.js --timeout=60000"
   },
   "dependencies": {
     "@google-cloud/functions-framework": "^3.0.0"
   },
   "devDependencies": {
     "mocha": "^9.1.3",
-    "p-retry": "^4.6.1",
     "sinon": "^13.0.0",
-    "supertest": "^6.0.0",
-    "uuid": "^8.3.2"
+    "supertest": "^6.0.0"
   }
 }
